# api/docs/swagger.yaml
definitions:
...
  models.Drink:
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/menu/dishes/query:
    get:
      description: Goes to the image/menu/... directory and sends the file.
      parameters:
      - description: spice
        in: path
        name: query
        required: true
        type: integer
      - description: course
        in: path
        name: query
        required: true
        type: integer
      - description: allergens
        in: path
        name: query
        required: true
        type: string
      - description: calories
        in: path
        name: query
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Get the static images for the file directory
      tags:
      - filter
  /api/v1/menu/drinks:
    get:
      description: Returns the json of drinks
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Lists all the drinks
      tags:
      - drinks
    post:
      description: Added a drink to the database regardless if it alread exists or
        not
      parameters:
      - description: JSON object of an order to be added or updated
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Drink'
      produces:
      - text/plain
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Posts a drink
      tags:
      - drinks
    put:
      description: Puts an drink, if it does not exist it is added otherwise it updates
        the order of the same ID
      parameters:
      - description: JSON object of an order to be added or updated
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Drink'
      produces:
      - text/plain
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Put a drink
      tags:
      - drinks
  /api/v1/menu/drinks.{id}:
    put:
      description: Gets a drink, using the primary key
      parameters:
      - description: ID of the dish
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Gets a drink
      tags:
      - drinks
  /api/v1/menu/drinks/{id}:
    delete:
      description: Deletes a drink using it's ID
      parameters:
      - description: ID of the dish
        in: path
        name: id
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Deletes a drink
      tags:
      - drinks
    patch:
      description: Updates the drink at the certain ID
      parameters:
      - description: JSON object of an order to be added or updated
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Drink'
      produces:
      - text/plain
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Updates a drink
      tags:
      - drinks
...
  /images/{dir}/{file}:
    get:
      description: Goes to the image/menu/... directory and sends the file.
      parameters:
      - description: dishes or drinks
        format: string
        in: path
        name: dir
        required: true
        type: string
      - description: name of the file
        format: string
        in: path
        name: file
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
      summary: Get the static images for the file directory
      tags:
      - image
    post:
      description: Takes the file and saves it into the path
      parameters:
      - description: dishes or drinks
        format: string
        in: path
        name: dir
        required: true
        type: string
      - description: name of the file
        format: string
        in: path
        name: file
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
      summary: Posts the image into the file directory
      tags:
      - image
swagger: "2.0"